requirements.txt
streamlit==1.32.0
selenium==4.18.1
beautifulsoup4==4.12.3
python-docx==1.1.0
groq==0.4.2
python-dotenv==1.0.1
Pillow==10.2.0
webdriver-manager==4.0.1
requests==2.31.0
google-generativeai==0.3.2
stability-sdk==0.8.5
regex==2023.6.3
lxml==4.9.3
concurrent-log-handler==0.9.24
nest-asyncio==1.5.8
chromedriver-autoinstaller==0.6.2 